Move Debian.yml vars to Ubuntu.yml
I was using ansible_os_family to get settings for Debian-family
hosts, but this doesn't work so well when you have an apt_mirror
which is only Debian or Ubuntu, for example.

I don't have any Debian hosts here, but anyways, it's better this
way so I can be more flexible in the future.

Signed-off-by: Alan Orth <alan.orth@gmail.com>